fix(platform): dynamically increase latency on underrun
After installing Pulse Audio, ALSA began to underrun like crazy. As a
workaround, the audio latency has been been bundled into the ring_buffer
struct and is dynamically increased (+10%) whenever underrun is
detected.

Also changed alsa to find a periods setting near 2, as this it is not a
valid value if Pulse Audio is installed.